backend/requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.13.3
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
cryptography==44.0.1
dataclasses-json==0.6.7
datasets==3.5.0
decorator==5.2.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docling==2.25.0
docling-core==2.21.1
docling-ibm-models==3.4.1
docling-parse==3.4.0
easyocr==1.7.2
ecdsa==0.19.0
email_validator==2.2.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evaluate==0.4.3
executing==2.2.0
fastapi==0.115.8
fastembed==0.6.0
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.12.0
gevent==24.11.1
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-docling==0.2.0
langchain-ollama==0.2.3
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.6
langdetect==1.0.9
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.10
latex2mathml==3.77.0
lazy_loader==0.4
loguru==0.7.3
lxml==5.3.1
Mako==1.3.9
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.4.7
onnxruntime==1.20.1
openai==1.65.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
portalocker==2.10.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyarrow==19.0.1
pyasn1==0.4.8
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
pypdf==5.3.0
pypdfium2==4.30.1
pyreadline3==3.5.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.1
pywin32==308
PyYAML==6.0.2
qdrant-client==1.13.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
Rtree==1.3.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semantic-router==0.0.72
semchunk==2.2.2
sentence-transformers==3.4.1
setfit==1.1.2
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlmodel==0.0.23
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.23
unstructured-client==0.30.3
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
websocket==0.2.1
websockets==15.0
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.2
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
